Add integration and unit tests for secret generation and masking commands
- Implemented comprehensive integration tests for the `secret` command group, covering various formats and options for secret generation and masking.
- Added unit tests for the `generate_secret` and `mask_secret` functions, ensuring correct behavior across different formats, lengths, and edge cases.
- Included tests for JSON and text output formats, case insensitivity, and validation of parameters.
- Ensured randomness in generated secrets and proper handling of special cases in masking.
