--- tests/strata/.strata/cli.yaml
@@ -1,3 +1,0 @@
-values:
-  last_execution_id: 89396b2d-2761-4fd2-a3b6-3bf7faa842eb
-  last_execution_on: '2026-05-31T11:46:07.837293+00:00'
