--- tests/strata/commands/secret/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for the secret command group."""
