+++ tests/strata/commands/test_commands_secret.py
